chatbot/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
httpx
pydantic
python-dotenv
